refactor: move embed_build_source to the engine (ADR-061 Phase 3)
Phase 3's last blocker of this shape. embed_build_source was a real engine
operation living in the CLI layer and raising Click exceptions, which
service_input_resolution then caught and translated -- the engine depending
on the CLI's error *type*.

The dependency closure was smaller than the ADR's note estimated. Of the
eight helpers it listed, four (_combine_packs, _filter_pack_layers,
_layer_value, route_inline_source_supplier) were already engine-side in
merge_support.py and needed nothing; is_inputs_pack_dir had already moved in
the pack-shape slice. The remaining three moved:

- buildsource/pack_load.py    -- the two pack loaders, raising SnapshotError
- buildsource/snapshot_exports.py -- exported_symbols_from_snapshot
- buildsource/embed.py        -- embed_build_source itself

The error contract was the actual work, and it is preserved exactly rather
than tidied. Two classes leave the engine and they mean different things:
ValidationError for a malformed .abicheck.yml (a usage error -> the CLI
renders click.UsageError -> cli.main remaps to exit 64) and SnapshotError for
an invalid pack (operational -> ClickException -> exit 1; the invocation was
well-formed, the data was not). Collapsing them would tell a CI consumer the
invocation was wrong when the data was. cli_buildsource.embed_build_source is
now a thin adapter doing exactly that translation, and keeping `quiet` as the
CLI's spelling of "hand the engine a stderr writer, or don't" -- an engine
module owns no stream, so warnings leave through an on_warning callback.

The typed surface still flattens both onto SnapshotError. That is not
tidiness either: widening what a Tier-2 caller must catch would be a breaking
API change made in passing.

Seven characterization tests were committed BEFORE the move (c027c43) and
pass unchanged after it. Every exit code was measured against the real CLI,
not read off the source, and re-measured after: corrupt pack still exits 1
with a byte-identical message, bad config still exits 64.

Test fallout was the ownership kind ADR-061 Context #4 names, not breakage:
23 tests patched or imported through the CLI facade. Engine-path patches now
target buildsource.embed; nine imports of _combine_packs now come from
merge_support, its real owner. One test's premise changed rather than its
target -- TestMalformedPackIsTranslated existed to assert the
ClickException->SnapshotError translation that no longer happens -- so it is
rewritten as TestMalformedPackErrorsReachTheTypedContract, asserting the
caller-visible contract (both classes arrive as SnapshotError) instead of the
mechanism, and moved next to the other error-contract tests.

Result: service_input_resolution.py has no CLI imports at all and is
classified `workflows`, the second service pipeline to get an owner. Two more
engine-CLI boundary allowlist entries are gone (15 -> 10 across this phase).
cli_buildsource.py drops 786 -> 576 lines as a side effect, not as the goal.
