+++ scripts/engine_cli_boundary.py
@@ -125,8 +125,6 @@ ENGINE_CLI_BOUNDARY_ALLOWLIST: frozenset[str] = frozenset(
         "abicheck/scan_engine.py::from .cli_scan_baseline import ...::1",
         "abicheck/scan_engine.py::from .cli_scan_helpers import ...::1",
         "abicheck/service_compare_pipeline.py::from .cli_buildsource import ...::1",
-        "abicheck/service_input_resolution.py::import click::1",
-        "abicheck/service_input_resolution.py::from .cli_buildsource import ...::1",
         "abicheck/service_scan.py::import click::1",
         "abicheck/service_scan.py::from .cli_scan_baseline import ...::1",
         "abicheck/service_scan.py::from .cli_scan_receipt import ...::1",
